Add support for using wxWebViewChromium from non-main thread
In this case -- which unavoidably happens when embedding this class in
Erlang/OTP, for example -- Chromium creates its own GMainContext and we
need to manually dispatch the events for it because the standard message
loop doesn't do it and if we don't do it neither, the browser window
doesn't do anything at all, including repainting itself.

Interestingly, this allows to simplify browser shutdown logic as we
don't need to dispatch wx events any longer, just the X/GTK ones, to
allow Chromium to finish its destruction.