Update binaries.md mentioning the build subdirectory and the location of wxwidgets.props
This very important document says 'You should have only include and lib subdirectories under it, nothing else'. This is not true. You also have a build subdirectory, (just as well as without it nothing further will work), and the vital wxwidgets.props file. The next paragraph says 'simply add the provided wxwidgets.props file...' and this was initially mysterious to me as I messed things up the first time and the instructions did not give a clue as to where the file might be and in fact misled me as they said the location where it should be was supposed to be empty bar two subdirectories.