Remove wxTextCtrl frame with wxBORDER_NONE for GTK3
This was deleted for GTK3 in 80e2264373 (Fix wxBORDER_NONE handling for
wxTextCtrl under GTK+ 3, 2017-10-31), but there was no harm in keeping it,
and it actually helps with a transitory visual glitch when the widget is
created on a parent that is already showing.

See #576, #24105.

(cherry picked from commit 2ddfcc2b2830bafa096b4830bf915c1cdefeb816)