Add missing default constructor of wxFileDialog in wxMSW
This ctor exists in the other ports and there is no reason for it not to
exist in wxMSW too.

This is a combined cherry pick of 11ed91af8f (Add missing default
constructor of wxFileDialog in wxMSW, 2023-12-03) and f8a6f0cd39
(Initialize wxFileDialog::m_data in default ctor in wxMSW too,
2023-12-18) from master.

See #24113.