Fix event dispatching from secondary thread in wxGTK
When Chrome creates a thread-specific GMainContext, as it does whenever
it's initialized from any thread other than the main process one, we
need to ensure that events are dispatched for this context whenever the
main event loop wakes up, as otherwise it could wake up, realize that
there is nothing for it to do, and resume polling again, without giving
us a chance to dispatch the events meant for this context.

To do this, register a special dummy event source which does nothing
except calling wxWakeUpIdle() whenever its "check" function is called,
as it happens when any events are available.