Fix crash when processing renames in wxMSW wxFileSystemWatcher
It could (and did, albeit rarely) happen that we read only the first of
the couple events associated with a file rename and in this case the
code crashed due to dereferencing invalid iterator.

Ideal would be to delay sending the "renamed" event until we get the new
path, but for now just fix the crash by not using the invalid iterator.

See #24119.

(cherry picked from commit e758d8ad80a83fe30e6ff2bae0e15653e7c32402)