Use narrower layout for wxGenericTreeCtrl in all cases
Indent and spacing values were reduced back in 618a5e382a (Applied
Greg's patch to tree control and related., 2001-08-18) compared to the
defaults of 15 and 18 pixels respectively, but only when not using
wxTR_NO_LINES style.

There doesn't seem to be any good reason to for these values to depend
on the style, so use the smaller values always, as they look better and
notably don't result in a lot of blank space between the branch toggle
icon and the image when using images.