Recommend building just "minimal" sample
Don't tell people to build all the samples and demos because this takes
time and some of them are badly outdated. Do tell them to build and run
the minimal one to check that things work as intended.

Also remove links to the wiki from wxOSX instructions, there is not much
useful from non-archaeological point of view there.