Fix wxAnyButton Current/Normal state under wxQt
Do not check for hasMouseTracking in QtGetCurrentState as mouse tracking
is always enabled in any widget derived from wxQtEventSignalHandler.

This resulted in an QtGetCurrentState reporting State_Current when it
should have reported State_Normal i.e. the button showed hovering state
when it was actually not hovering.

After the change, button state will show State_Current only if the mouse
pointer is actually hovered on the button.

Closes #24129.

Closes #24132.