Try to avoid bogus ABI break error after the last commit
Apparently abicheck detects removal of the dependency on a weak symbol
as an ABI breakage because this symbol is not exported from wxPropGrid
library any more. This has no effect in practice because any program
linking with this library also links with wxBase, which still has it,
but try to avoid it nevertheless just to keep the ability to detect real
ABI breaks in the future.

For the record, the error given by abicheck for libwx_gtk3u_propgrid-3.2
was:

Functions changes summary: 1 Removed, 0 Changed, 0 Added function
Variables changes summary: 0 Removed, 0 Changed, 0 Added variable

1 Removed function:

  [D] 'method wxUniChar::wxUniChar(const wxUniCharRef&)'    {_ZN9wxUniCharC2ERK12wxUniCharRef@@WXU_3.2, aliases _ZN9wxUniCharC1ERK12wxUniCharRef@@WXU_3.2}