Fix close button handling in wxAuiTabContainer::IsTabVisible()
The existing code didn't work correctly when wxAUI_NB_CLOSE_ON_ALL_TABS
style was used as it didn't check for it and so assumed that the close
button was only shown on the current page.

This resulted in wrong result being returned from this function for the
right most tab which could not actually fit into the available space
when rendered (because Render() did draw all close buttons correctly)
and, because of this, not showing the tab entirely when MakeTabVisible()
was called, e.g. when adding a new tab.

Closes #24096.