Don't define a separate wxAuiManager::OnHintActivate()
This was only used under Mac but defined under all platforms.

Don't bother with defining it elsewhere and just use a trivial empty
lambda as wxEVT_ACTIVATE handler under Mac.

Note that it's not really clear if this is still needed neither, but it
doesn't cost much to keep this workaround just in case it is.