Avoid resizing wxAuiToolBar to size in the wrong orientation
This is unnecessary and resulted in sometimes visible flicker, when the
wrong shape of the toolbar could be temporarily seen on screen before it
was replaced with the correct one.

Only set the actual window size in Realize(), and do it only for the
current orientation, instead of doing it in RealizeHelper() which is
called for both the current orientation and the transversal one.

Closes #24023.