Update new page in AUI tab control after removing the active one
Previously this was only done when removing the global (i.e. per
notebook) current page, but not when removing the page which was current
in this tab control without being globally current, and this resulted in
not showing the newly shown page correctly.

Fix this by explicitly updating the tab control in this case.

Closes #24063.