Add wxAuiTabCtrl::DoShowTab() helper
No real changes, just add a function ensuring that the tab is shown.

It's not really clear why the existing code sometimes calls
MakeTabVisible() after SetActivePage() and sometimes doesn't, and
perhaps it should actually always do it, in which case this function
could be merged with SetActivePage() to do everything at once, but for
now keep the changes minimal.