Ensure full label is shown after editing it in generic wxListCtrl
The list control needs to be updated after the user has edited an item
label because, for example, in icon view the column width and positions
of subsequent columns may change if the edited label is or was wider
than the icons and other labels in the column.

See #24311.

(cherry picked from commit 716bf59449521051c419e3d0f977161b97dcc5f7)