Stop checking if GDK window is our own in GTK callbacks
GTKIsOwnWindow() doesn't always return the correct result, i.e. it
doesn't recognize all windows that a GTK widget is made of as being
"ours", so checking for this resulted in wrongly ignoring some events.

And this check doesn't seem to be actually necessary, as we shouldn't be
ever called by GTK for the windows that are not ours, so just remove it.

See #24245.