Generate wxEVT_ENTER_WINDOW in icon part of wxSearchCtrl in wxGTK
Mouse motion events happening over GtkEntry icons are consumed by this
class "event" callback, which simply returns GDK_EVENT_STOP for them if
the icon is insensitive (which is the case for GtkSearchEntry) and so
our "enter-notify-event" callback is never called.

Fix this by handling "event" signal ourselves and generating
wxEVT_ENTER_WINDOW from it if necessary.

Doing this required adding a global pointer to the window currently
under mouse, which is also used for generating wxEVT_LEAVE_WINDOW if
necessary.

See #24245.