Add a hack to avoid a problem with missing wxStaticText text
Somehow, for some combination of characters, the native control didn't
show anything at all in the control unless its size was one bigger than
the size we computed. E.g. adding the following code to minimal sample:

    auto p = new wxPanel(this);
    new wxStaticText(p, wxID_ANY, " AI", wxPoint(10, 10));
    new wxStaticText(p, wxID_ANY, " AJ", wxPoint(10, 40));
    new wxStaticText(p, wxID_ANY, " AK", wxPoint(10, 70));

    new wxStaticText(p, wxID_ANY, " BI", wxPoint(100, 10));
    new wxStaticText(p, wxID_ANY, " BJ", wxPoint(100, 40));
    new wxStaticText(p, wxID_ANY, " BK", wxPoint(100, 70));

showed just a "hole" instead of the " AJ" string, even though all the
others appeared correctly (observed under both Windows 10 and 11, at
standard and 200% DPI).

This looks like a bug both in our code (because we seem to not be
computing the text extent correctly) and in the native control (because
it should still show at least something even if the last pixel doesn't
fit instead of showing nothing at all), but it's not clear how to fix
the former and there is nothing we can do about the latter, so add an
extra pixel to fix a pretty bad user-visible problem.