BUGFIX: On wxAuiToolbar tools that do not fit are shown bunched up together at the end
The 3bb4e68f02ca74d530a8c4a875119cd16a87e949 commit solved one problem introducing another;
if a tool did not fit in the toolbar it was rendered at the end of the toolbar nevertheless,
because it used the size computed by the sizer to check for fit but that size could become
zero once the tools exausted all the available space in the toolbar.
This fix used the dimensions provided by the art provider for the fit check.