Prefer EGL configurations without alpha if it wasn't requested
At least Mesa Iris on Wayland is known to put RGBA5551 and RGBA4444
formats at the start of the config list, even if we didn't request
alpha, which is unexpected, so filter them out manually in this case to
ensure we get the expected RGB444, as with the other drivers.

See #24395.

Co-authored-by: Vadim Zeitlin <vadim@wxwidgets.org>