Fix getting POSIX tag for locales such as BCP47 zh-Hans-CN
Previously GetTag() returned "zh_CN@" for this locale because "Hans" is
not recognized as script alias for some reason. It's not clear whether
this is correct or not, but we shouldn't use empty script value in any
case.

This incidentally fixes language lookup under macOS when this is the
preferred user language, as now it matches "zh_CN" when looking for the
best available translation in wxTranslations, whereas previously it did
not.

See #24396.