Revert disk clean up in GitHub Actions CI workflow
This reverts the actual change of db65742dbc (Remove some files on
GitHub Action runner to free up some space, 2024-03-16), while keeping
the "df" command just in case, as we were not really running out of disk
space and the error messages about it were a red herring.

The real problem will be fixed by the next commit.