Simplify wxrc code by removing broken support for multiple bases
This code was added back in aac18ec71e (generate ctors with optional
parent parameters in C++ code (patch 1238355), 2006-05-06) but
apparently never worked, as it generated uncompilable code if it was
executed (see #24383) and doesn't seem to make any sense as the name of
the class defined in XRC can never be "wxMenu" or any other of the
already existing wx classes.

Revert this to simplify the code.