Fix scale computation in wxBitmapBundle for some bitmap sizes
GetNextAvailableScale() returned wrong result when the bitmap size was
not exactly divisible by the scale, but this didn't matter for square
bitmaps because multiplying the default size by the not quite correct
scale still rounded to the correct size.

This was not the case for non-square bitmaps, however, as all our
computations use only bitmap height and dividing bitmap width by the
exact scale and multiplying it by the scale computed from the scaled and
unscaled height values could result in a value different from the
initial one, as shown by the newly added unit test, which failed before
this commit.

Fix this by computing the ratio of the bitmap to the base one and then
multiplying this ratio by the bitmap scale.

Closes #24433.