Fix handling of sizes differing in width only in wxBitmapBundle
We assumed that bitmaps with the same height also have the same width,
but this is not necessarily true and if it did happen, GetBitmap()
failed it postcondition of returning the bitmap with the requested size.

Fix this by rescaling the closest bitmap to have the correct width even
if it already has the correct height.

Also enhance the bitmap comparator used for sorting to work correctly
with the bitmaps of the same height but different width.

Closes #24434.