Fix calling wxOverlay::Reset() without Init() first under Mac
It is valid to call Reset() even if Init() had never been called, e.g.
when we didn't need to draw anything on the overlay, but it asserted and
crashed in this case because it used invalid m_overlayWindow under Mac.

Fix this by simply doing nothing in Reset() if m_overlayWindow was not
initialized.

This commit is best viewed ignoring whitespace-only changes.

Closes #24415.

Co-Authored-By: Daniel Kulp <dan@kulp.com>