wxrc: `FileToCppArray`: Byte-array initialization integer literals are now hex; progress is shown.
16 bytes are output on each line.

Previously, a variable number of bytes used to be output on each line,
depending on the character length of the decimal representation of each bytes.
Indeed, the number of output characters used to be lower;
the increase is around 20%; but the representation is nicer for some people (I hope).

Progress (offset, total size, percentage) is shown every 256 bytes and at the end (as comments).

Example of output (generated C++ source code):

```
// Temporary file    7h of   C9h (  3.48%) (of size       68h) '/home/ubuntu/Projects/com_github/wxWidgets/wxWidgets-build/Example2.cpp$ExampleImages_Example_0007h.png':
const size_t xml_res_size_7 = 104;
const unsigned char xml_res_file_7[] = {
// Offset        0h of       68h (   0.00%):
0x89,0x50,0x4e,0x47, 0x0d,0x0a,0x1a,0x0a, 0x00,0x00,0x00,0x0d, 0x49,0x48,0x44,0x52,
0x00,0x00,0x00,0x10, 0x00,0x00,0x00,0x10, 0x08,0x02,0x00,0x00, 0x00,0x90,0x91,0x68,
0x36,0x00,0x00,0x00, 0x09,0x70,0x48,0x59, 0x73,0x00,0x00,0x0e, 0xc4,0x00,0x00,0x0e,
0xc4,0x01,0x95,0x2b, 0x0e,0x1b,0x00,0x00, 0x00,0x1a,0x49,0x44, 0x41,0x54,0x28,0xcf,
0x63,0x6c,0x60,0xf8, 0xcf,0x40,0x0a,0x60, 0x62,0x20,0x11,0x8c, 0x6a,0x18,0xd5,0x30,
0x74,0x34,0x00,0x00, 0xc5,0xbf,0x01,0x9f, 0x22,0x91,0xff,0xbd, 0x00,0x00,0x00,0x00,
0x49,0x45,0x4e,0x44, 0xae,0x42,0x60,0x82
// Offset       68h of       68h ( 100.00%).
};
```