Add filters support to wxOSX wxFileDialog::ShowWindowModal() too
This was already implemented in ShowModal() but not in this function, so
just copy the code to it too.

It would, of course, be better to refactor this class to avoid
duplicating all this code, but for now this adds this useful feature at
least in this way.

See #24438.

(cherry picked from commit 5e22bc3f3c240f51524376f4d430208f6dd856e3)