Support building universal CEF wrapper and wxwebview library
Download all (usually just two) CEF versions corresponding to the
architectures in CMAKE_OSX_ARCHITECTURES if the latter is defined and
lipo them together to allow building universal binaries of
libcef_dll_wrapper and wxwebview library using it.

This makes it much easier to build universal applications using CEF
wxWebView backend as they can now be built in exactly the same way as
usual and just defining CMAKE_OSX_ARCHITECTURES appropriately, instead
of having to manually lipo everything together in the application
makefiles.

As a side effect, this simplifies cef_version_info.cmake by moving all
not version-specific logic from it to the main makefile itself and
leaving just the variables needed to construct the CEF download path and
its checksums in this file.