Switch to direct wxConfig creation in the sample
The claimed advantages of cresting it on demand don't really apply to
the sample, so don't do it like this, but create the object directly to
make it easier to customize the creation flags.

Still make it possible to test the on-demand creation by predefining
WX_CONFIG_CREATE_ON_DEMAND environment variable, this is simpler than
telling people to uncomment some lines in the sample code.