Stop inheriting picker buttons from wxButton in wxGTK
This resulted in many problems, such as the previously fixed (using more
or less ugly workarounds) #8107 and #17984 and another similar problem
which happened when disabling a picker control, all due to the fact that
GtkFileChooserButton does not inherit from GtkButton at GTK level, but
wxButton, rather reasonably, assumes that its widget is a GtkButton.

Solve this once and for all by not using inheritance wrongly here and
using composition instead: now wxFileButton contains wxGenericFileButton
instead of inheriting from it (and the same thing for wxDirButton) and
just forwards its functions to it in this case.

The only new complication is that wxFileDirPickerWidgetBase::GetPath()
had to be made virtual too, to allow retrieving the correct path when
the generic button is used, but otherwise the new code is much simpler,
in addition to being more correct, and, notably, the type unsafe
UpdateDialogPath() and UpdatePathFromDialog() functions could be
entirely removed and many other casts eliminated.

For completeness, notice that there is an alternative solution for the
original problem: we could use composition at wxGenericFileButton level
and derive it from wxControl and make it contain a wxButton and then
wxGTK wxFileButton could continue inheriting from it. However it seems
more natural to let wxGenericFileButton be a wxButton that it really is
and so this solution was chosen instead.