Freeze wxPropertyGrid while populating it from XRC
Call wxPropertyGridPopulator::SetGrid() to associate the grid with the
"populator" and let it freeze and thaw it as needed, to avoid too many
unnecessary intermediate updates.

The effect doesn't seem to be dramatic, but doing this can only help, so
do it just in case it helps more with the bigger grids.