Stop auto-centering splitter after it was moved by user
This didn't happen before due to the check for "FromEvent" added to
DoSetSplitter() back in f525476870 (Added wxPropertyGridManager header
support; Refactored wxPropertyGrid DoSetSplitterPosition() code,
2009-12-13) without any real explanation in the commit message and which
appears to be wrong and, at the very list, inconsistent with
wxPropertyGrid::DoSetSplitter() which sets m_dontCenterSplitter to true
if "FromEvent" is present in the flags.

Simply remove the check for it here and only avoid setting
m_isSplitterPreSet when auto-centering it from wxPropertyGrid own code.